Broken

+++ src/Core/Classes/Resources/Buffer.hpp
@@ -21,7 +21,7 @@ namespace nihil
         Resource() {Logger::Log("The specialized constructor for vk::Buffer called.");};
         void destroy() override
         {
-            if(destroyed) return;
+            if(destroyed || !assigned) return;
             device.destroyBuffer(res);
             destroyed = true;
